Primeiro Commit para Novo Nome

+++ socios/views.py
@@ -6,15 +6,19 @@ from django.views.generic import  UpdateView, DeleteView # ListView, CreateView
 from django.urls import reverse_lazy
 from django.contrib import messages
 
+from django.conf import settings
+LINHAS_POR_PAGINA = settings.LINHAS_POR_PAGINA
 
-
+#niveis_permitidos = ['admin', 'financeiro', 'social', 'esportivo', 'livre'] 
 #====================================================================================
 def lista_mensalidades(request):
+
     query = request.GET.get('q') 
-    socios = Socio.objects.prefetch_related('mensalidades').all()  # Agora usa 'dependentes'
+    #socios = Socio.objects.prefetch_related('mensalidades').all()  
+    socios = Socio.objects.filter(tipo_socio__in=['patrimonial', 'contribuinte']).prefetch_related('mensalidades')
     if query: 
        socios = socios.filter(nome__icontains=query) 
-    paginator = Paginator(socios, 10)  # Display 10 items per page
+    paginator = Paginator(socios, LINHAS_POR_PAGINA)  # Display 10 items per page
     page_number = request.GET.get('page')
     page_obj = paginator.get_page(page_number)
     return render(request, 'socios/lista_mensalidades.html', {'socios':page_obj.object_list, 'page_obj': page_obj})
@@ -25,12 +29,35 @@ class MensalidadeUpdateView(UpdateView):
     form_class = MensalidadeForm
     template_name = 'socios/mensalidade_form.html'
     success_url = reverse_lazy('gestao_mensalidades')
-
+    #####  ROTINA DE PERMISSOES  #######
+    def dispatch(self, request, *args, **kwargs):
+        if not request.user.is_authenticated:
+            messages.error(request, "Você não está logado.")
+            return redirect('home')  
+        niveis_permitidos = ['admin', 'financeiro', 'social']
+        if not hasattr(request.user, 'nivel_acesso') or request.user.nivel_acesso not in niveis_permitidos:
+            messages.warning(request, "Você não tem permissão para acessar esta página.")
+            return redirect('home') 
+        # Se o usuário tiver permissão, continua com a execução normal da view
+        return super().dispatch(request, *args, **kwargs)
+    #########################################
 #====================================================================================
 class MensalidadeDeleteView(DeleteView):
     model = Mensalidade
     template_name = 'socios/mensalidade_confirm_delete.html'
     success_url = reverse_lazy('gestao_mensalidades')
+    #####  ROTINA DE PERMISSOES  #######
+    def dispatch(self, request, *args, **kwargs):
+        if not request.user.is_authenticated:
+            messages.error(request, "Você não está logado.")
+            return redirect('home')  
+        niveis_permitidos = ['admin'] 
+        if not hasattr(request.user, 'nivel_acesso') or request.user.nivel_acesso not in niveis_permitidos:
+            messages.warning(request, "Você não tem permissão para acessar esta página.")
+            return redirect('home') 
+        # Se o usuário tiver permissão, continua com a execução normal da view
+        return super().dispatch(request, *args, **kwargs)
+    #########################################
 
 #==============================================================
 def home(request):
@@ -44,12 +71,11 @@ def sobre(request):
 def gerenciamento(request):
     query = request.GET.get('q') 
     active_tab = request.GET.get('activeTab', 'tab1')  # Valor padrão é 'tab1'
-    print(active_tab)
     socios = Socio.objects.all()
     if query: 
         socios = Socio.objects.filter(nome__icontains=query) 
 
-    paginator = Paginator(socios, 10)  # Display 10 items per page
+    paginator = Paginator(socios, LINHAS_POR_PAGINA)  # Display 10 items per page
     page_number = request.GET.get('page')
     page_obj = paginator.get_page(page_number)
     dependentes = Dependente.objects.all()
@@ -62,7 +88,7 @@ def gerenciamento(request):
             if query: 
                 socios = Socio.objects.filter(nome__icontains=query) 
 
-            paginator = Paginator(socios, 10)  # Display 10 items per page
+            paginator = Paginator(socios, LINHAS_POR_PAGINA)  # Display 10 items per page
             page_number = request.GET.get('page')
             page_obj = paginator.get_page(page_number)
             dependentes = Dependente.objects.all()
@@ -71,7 +97,7 @@ def gerenciamento(request):
 
         elif active_tab == 'tab2':
             dependentes = Dependente.objects.all()
-            paginator = Paginator(dependentes, 10)  # Display 10 items per page
+            paginator = Paginator(dependentes, LINHAS_POR_PAGINA)  # Display 10 items per page
             page_number = request.GET.get('page')
             page_obj = paginator.get_page(page_number)
             socios = Socio.objects.all()
@@ -80,7 +106,7 @@ def gerenciamento(request):
 
         elif active_tab == 'tab3':
             mensalidades = Mensalidade.objects.all()
-            paginator = Paginator(mensalidades, 10)  # Display 10 items per page
+            paginator = Paginator(mensalidades, LINHAS_POR_PAGINA)  # Display 10 items per page
             page_number = request.GET.get('page')
             page_obj = paginator.get_page(page_number)
             socios = Socio.objects.all()
@@ -92,12 +118,14 @@ def gerenciamento(request):
 #====================================================================================
 def gestao_socios(request):
     query = request.GET.get('q') 
-    socios = Socio.objects.all()
+    #socios = Socio.objects.all()  # Retorna so os ativos = True
+    socios = Socio._base_manager.all() # Retorna todo ativos e inativos = True e False
     if query: 
         socios = socios.filter(nome__icontains=query) 
 
-    paginator = Paginator(socios, 10)  # Display 10 items per page
+    paginator = Paginator(socios, LINHAS_POR_PAGINA)  # Display 10 items per page
     page_number = request.GET.get('page')
+
     page_obj = paginator.get_page(page_number)
     context = {'socios':page_obj.object_list, 'page_obj': page_obj}
     
@@ -110,7 +138,7 @@ def gestao_dependentes(request):
     if query: 
         dependentes = dependentes.filter(nome__icontains=query) 
 
-    paginator = Paginator(dependentes, 10)  # Display 10 items per page
+    paginator = Paginator(dependentes, LINHAS_POR_PAGINA)  # Display 10 items per page
     page_number = request.GET.get('page')
     page_obj = paginator.get_page(page_number)
     context = {'dependentes':page_obj.object_list, 'page_obj': page_obj}
@@ -124,7 +152,7 @@ def gestao_mensalidades(request):
     if query: 
         mensalidades = mensalidades.filter(socio__nome__icontains=query)
 
-    paginator = Paginator(mensalidades, 10)  # Display 10 items per page
+    paginator = Paginator(mensalidades, LINHAS_POR_PAGINA - 4)  # Display 10 items per page
     page_number = request.GET.get('page')
     page_obj = paginator.get_page(page_number)
     context = {'mensalidades':page_obj.object_list, 'page_obj': page_obj}
@@ -146,7 +174,7 @@ def lista_socio_dependentes(request):
     socios = Socio.objects.prefetch_related('dependentes').all()  # Agora usa 'dependentes'
     if query: 
         socios = socios.filter(nome__icontains=query) 
-    paginator = Paginator(socios, 10)  # Display 10 items per page
+    paginator = Paginator(socios, LINHAS_POR_PAGINA)  # Display 10 items per page
     page_number = request.GET.get('page')
     page_obj = paginator.get_page(page_number)
     context = {'socios':page_obj.object_list, 'page_obj': page_obj}
@@ -159,11 +187,21 @@ def lista_socios(request):
 
 #==============================================================
 def cria_socio(request):
+    ### ROTINA DE PERMISSOES  ###
+    if not request.user.is_authenticated:
+        messages.error(request, "Você não está logado.")
+        return redirect('home')  
+    niveis_permitidos = ['admin', 'financeiro', 'social'] 
+    if not hasattr(request.user, 'nivel_acesso') or request.user.nivel_acesso not in niveis_permitidos:
+        messages.warning(request, "Você não tem permissão para acessar esta página.")
+        return redirect('home') 
+    ###############################
+
     if request.method == 'POST':
         form = SocioForm(request.POST, request.FILES)
         if form.is_valid():
             form.save()
-            messages.success(request, f'Socio incluido com sucesso.')
+            messages.success(request, 'Socio incluido com sucesso.')
             return redirect('gestao_socios')
         else:
             messages.error(request, 'Ocorreu um erro ao processar o formulário.')
@@ -173,13 +211,23 @@ def cria_socio(request):
 
 #==============================================================
 def edita_socio(request, pk):
-    socio = get_object_or_404(Socio, pk=pk)
-
+    ### ROTINA DE PERMISSOES  ###
+    if not request.user.is_authenticated:
+        messages.error(request, "Você não está logado.")
+        return redirect('home')  
+    niveis_permitidos = ['admin', 'financeiro', 'social'] 
+    if not hasattr(request.user, 'nivel_acesso') or request.user.nivel_acesso not in niveis_permitidos:
+        messages.warning(request, "Você não tem permissão para acessar esta página.")
+        return redirect('home') 
+    ###############################
+    #socios = Socio.objects.all()  # Retorna so os ativos = True
+    socios = Socio._base_manager.all() # Retorna todo ativos e inativos = True e False
+    socio = get_object_or_404(socios, pk=pk)
     if request.method == 'POST':
         form = SocioFormEdit(request.POST, request.FILES, instance=socio)
         if form.is_valid():
             form.save()
-            messages.success(request, f'Socio alterado com sucesso.')
+            messages.success(request, 'Socio alterado com sucesso.')
             return redirect('gestao_socios')
         else:
             messages.error(request, 'Ocorreu um erro ao processar o formulário.')
@@ -190,6 +238,15 @@ def edita_socio(request, pk):
 
 #==============================================================
 def exclui_socio(request, pk):
+    ### ROTINA DE PERMISSOES  ###
+    if not request.user.is_authenticated:
+        messages.error(request, "Você não está logado.")
+        return redirect('home')  
+    niveis_permitidos = ['admin', 'financeiro', 'social'] 
+    if not hasattr(request.user, 'nivel_acesso') or request.user.nivel_acesso not in niveis_permitidos:
+        messages.warning(request, "Você não tem permissão para acessar esta página.")
+        return redirect('home') 
+    ###############################
     socio = get_object_or_404(Socio, pk=pk)
     if request.method == 'POST':
         socio.delete()
@@ -204,6 +261,16 @@ def lista_dependentes(request, socio_id):
 
 #==============================================================
 def cria_dependente(request):
+    ### ROTINA DE PERMISSOES  ###
+    if not request.user.is_authenticated:
+        messages.error(request, "Você não está logado.")
+        return redirect('home')  
+        #niveis_permitidos = ['admin', 'financeiro', 'social', 'esportivo', 'livre'] 
+    niveis_permitidos = ['admin', 'financeiro', 'social']
+    if not hasattr(request.user, 'nivel_acesso') or request.user.nivel_acesso not in niveis_permitidos:
+        messages.warning(request, "Você não tem permissão para acessar esta página.")
+        return redirect('home') 
+    ###############################
     if request.method == 'POST':
         form = DependenteForm(request.POST, request.FILES)  # Preenche o formulário com os dados enviados
         if form.is_valid():
@@ -216,6 +283,15 @@ def cria_dependente(request):
 
 #==============================================================
 def edita_dependente(request, dependente_id):
+    ### ROTINA DE PERMISSOES  ###
+    if not request.user.is_authenticated:
+        messages.error(request, "Você não está logado.")
+        return redirect('home')  
+    niveis_permitidos = ['admin', 'financeiro', 'social']
+    if not hasattr(request.user, 'nivel_acesso') or request.user.nivel_acesso not in niveis_permitidos:
+        messages.warning(request, "Você não tem permissão para acessar esta página.")
+        return redirect('home') 
+    ###############################
     dependente = get_object_or_404(Dependente, pk=dependente_id)  # Obtém o dependente pelo ID
     if request.method == 'POST':
         form = DependenteFormEdit(request.POST, request.FILES, instance=dependente)  # Preenche o formulário com os dados enviados
@@ -228,9 +304,21 @@ def edita_dependente(request, dependente_id):
 
 #======================================================================================================================
 def exclui_dependente(request, dependente_id):
+    ### ROTINA DE PERMISSOES  ###
+    niveis_permitidos = ['admin', 'financeiro', 'social']
+    if not request.user.is_authenticated:
+        messages.error(request, "Você não está logado.")
+        return redirect('home')  
+    if not hasattr(request.user, 'nivel_acesso') or request.user.nivel_acesso not in niveis_permitidos:
+        messages.warning(request, "Você não tem permissão para acessar esta página.")
+        return redirect('home') 
+    ###############################
+
     dependente = get_object_or_404(Dependente, pk=dependente_id) 
     if request.method == 'POST':
         dependente.delete()  # Exclui o dependente do banco de dados
         return redirect('gestao_dependentes')  # Redireciona para a lista de dependentes
     return render(request, 'socios/exclui_dependente.html', {'dependente': dependente})
+
+
